+++ src/obj.c
@@ -71,17 +71,19 @@ void _Nst_obj_free(Nst_Obj *obj)
 {
     assert(Nst_HAS_FLAG(obj, Nst_FLAG_OBJ_DESTROYED));
 
+    if (Nst_HAS_FLAG(obj, Nst_FLAG_GGC_PRESERVE_MEM))
+        return;
+
+    Nst_TypeObj *ob_t = obj->type;
+
     if (Nst_HAS_FLAG(obj, Nst_FLAG_GGC_IS_SUPPORTED)) {
         Nst_GGCObj *ggc_obj = GGC_OBJ(obj);
         Nst_GGCList *ls = ggc_obj->ggc_list;
 
         // if ls is NULL it means that the object is being deleted in a garbage
         // collection
-        if (ls == NULL) {
-            if (Nst_HAS_FLAG(obj, Nst_FLAG_GGC_DELETE))
-                goto free_mem;
-            return;
-        }
+        if (ls == NULL)
+            goto free_mem;
 
         if (ls->head == ggc_obj)
             ls->head = GGC_OBJ(ggc_obj->p_next);
@@ -99,12 +101,19 @@ void _Nst_obj_free(Nst_Obj *obj)
 free_mem:
     if (obj->type->p_len >= _Nst_P_LEN_MAX) {
         Nst_free(obj);
+
+        if (obj != OBJ(ob_t))
+            Nst_dec_ref(ob_t);
+
         return;
     }
 
     obj->p_next = obj->type->p_head;
     obj->type->p_head = obj;
     obj->type->p_len++;
+
+    if (obj != OBJ(ob_t))
+        Nst_dec_ref(ob_t);
 }
 
 Nst_Obj *_Nst_inc_ref(Nst_Obj *obj)
@@ -119,10 +128,7 @@ void _Nst_dec_ref(Nst_Obj *obj)
 
     assert(obj->ref_count >= 0); // The ref_count should nevere be below zero
     if (obj->ref_count <= 0) {
-        Nst_TypeObj *ob_t = obj->type;
         _Nst_obj_destroy(obj);
         _Nst_obj_free(obj);
-        if (obj != OBJ(ob_t))
-            Nst_dec_ref(ob_t);
     }
 }