+++ docs/changes.md
@@ -90,12 +90,14 @@ _General changes_:
 - added `arg_num` argument to `Nst_call_func`
 - renamed `Nst_run_func_context` to `Nst_run_paused_coroutine` and modified its arguments
 - removed `Nst_state_init` and `Nst_state_free`
+- renamed `Nst_ENDIANNESS` to `Nst_BYTEORDER`
 
 **Bug fixes**
 
 - fixed `Nst_obj_eq_c` and `Nst_obj_ne_c` leaving a dangling reference of the resulting boolean object
 - fixed type objects not being freed properly
 - fixed many constants not being destroyed at the end of the program
+- fixed a possible read-after-free bug in the garbage collector
 
 ---